+++ src/main/java/market/demo/domain/CartItem.java
@@ -2,7 +2,6 @@ package market.demo.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
-import lombok.Setter;
 
 import static jakarta.persistence.FetchType.LAZY;
